fix:Fixed flicking popover and button in modal
player들이 이모티콘을 spamming 할 때, 게임 시작 버튼이 눌리지 않고, catchAnswer 페이지로 접속하는 QR코드가 깜빡이는 문제 #150
