+++ app/modules/popoverAtom.tsx
@@ -1,0 +1,13 @@
+import { atom } from "jotai"
+
+export const anchorElAtom = atom<HTMLElement | null>(null);
+
+export const setAnchorEl = atom(
+    null,
+    (get, set, popstate : HTMLElement | null) => {
+        set(
+            anchorElAtom,
+            popstate
+        )
+    }
+)
