+++ package.json
@@ -13,6 +13,7 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@fontsource/roboto": "^5.0.8",
+    "@mui/icons-material": "^5.14.19",
     "@mui/material": "^5.14.17",
     "@mui/styled-engine-sc": "^6.0.0-alpha.5",
     "@react-three/fiber": "^8.15.12",
@@ -23,6 +24,7 @@
     "cookie": "^0.6.0",
     "jotai": "^2.5.1",
     "material": "^0.7.5",
+    "material-ui-popup-state": "^5.0.10",
     "next": "14.0.2",
     "next-client-cookies": "^1.1.0",
     "nextjs-current-url": "^1.0.3",
